Move Translation to specific module
